refactor: clean up old FastAPI logic, temporary utils
