--- backend/rotini/envs/ci.py
@@ -1,5 +1,0 @@
-DATABASE_USERNAME = "postgres"
-DATABASE_PASSWORD = "test"
-DATABASE_HOST = "localhost"
-DATABASE_PORT = 5431
-DATABASE_NAME = "postgres"
