--- backend/rotini/envs/local.py
@@ -1,5 +1,0 @@
-DATABASE_USERNAME = "postgres"
-DATABASE_PASSWORD = "test"
-DATABASE_HOST = "localhost"
-DATABASE_PORT = 5432
-DATABASE_NAME = "postgres"
