--- backend/rotini/envs/migrate.py
@@ -1,3 +1,0 @@
-from envs.local import *
-
-DATABASE_HOST = "localhost"
