--- backend/rotini/exceptions.py
@@ -1,4 +1,0 @@
-class DoesNotExist(Exception):
-    """
-    General purpose exception signalling a failure to find a database record.
-    """
